+++ src/Credfeto.ChangeLog/Constants/TemplateFile.cs
@@ -33,6 +33,7 @@ public static class TemplateFile
             .Append(FileConstants.Unreleased)
             .Append("]\n")
             .AppendSectionHeadings(language.SectionOrder)
+            .Append('\n')
             .Append("<!--\n")
             .Append(
                 "Releases that have at least been deployed to staging, BUT NOT necessarily released to live.  Changes should be moved from [Unreleased] into here as they are merged into the appropriate release branch\n"